Module oak_transparency_claims::intoto

This module contains structs for specifying in-toto statements. The structs in this module must be kept in sync with the structs defined in https://github.com/project-oak/transparent-release/blob/main/pkg/intoto/intoto.go.

Structs

Statement

This struct represents a generic statement that binds a predicate to a particular subject.

Subject

A software artifact identified by its name and a set of artifacts.

Constants

STATEMENT_INTOTO_V01

URI representing in-toto v01 statements. This is constant for all predicate types.

Type Aliases

DigestSet